Heuristics
Created custom heuristic that privileges states with empty cells.

Renamed new-node to constructor and change it to receive the heuristic calculator and auto calculate the heuristic of that node if exists.

Changed get-pieces to get-state-pieces.

Changed node-spawner to use the new constructor form.
Changed node print up to root to verify if heuristic is -1, meaning it was not given when creating the node, it will not print the value.